pyproject.toml
[project]
name = "barohi.sc"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.3",
    "fastapi>=0.115.11",
    "google-cloud-vision>=3.10.1",
    "httpx>=0.28.1",
    "matplotlib>=3.10.1",
    "numpy>=2.2.4",
    "openai>=1.66.5",
    "opencv-python>=4.11.0.86",
    "pillow>=11.1.0",
    "scikit-learn>=1.6.1",
    "shapely>=2.0.7",
    "textdistance>=4.6.3",
    "uvicorn>=0.34.0",
]
